storage: improve copyv allowing it receive source slice and dst index+offset
This change is helpful when we optimize rafs io/read stack, aiming at
make it capable of handling multiple chunks once, thus to reduce
times of reading blobcache and backend.

+++ storage/src/device.rs
@@ -179,7 +179,10 @@ impl FileReadWriteVolatile for RafsBioDevice<'_> {
             return self.fill_hole(bufs);
         }
 
-        self.dev.rw_layer.load().read(&self.bio, bufs, offset)
+        self.dev
+            .rw_layer
+            .load()
+            .read(&self.bio, bufs, offset as usize)
     }
 
     fn write_at_volatile(&mut self, slice: VolatileSlice, _offset: u64) -> Result<usize, Error> {
